requirements.txt
numpy==1.23.4
tensorflow==2.11.0
tensorflow-datasets==4.8.2
tensorflow-estimator==2.11.0
tensorflow-gan==2.1.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-metadata==1.12.0
tensorflow-probability==0.19.0
ninja==1.11.1
scipy==1.10.0
matplotlib==3.7.0
tqdm==4.64.1
tensorflow_datasets==4.8.2
ml_collections==0.1.1
jax==0.4.6
jaxlib==0.4.6
urllib3==1.26.14
# calculate_flops 
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
opencv-python
pyiqa
absl-py
ml_collections
# causal_conv1d==1.1.1